Testing for both TreeNode impls
